+++ server/studyTime/src/main/java/com/ipseg/studyTime/api/user/dto/userJoin/UserJoinResponse.java
@@ -1,0 +1,11 @@
+package com.ipseg.studyTime.api.user.dto.userJoin;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class UserJoinResponse {
+    private String id;
+    private String name;
+}
